requirement.txt
amqp==5.1.1
asgiref==3.6.0
azure-core==1.26.4
azure-storage-blob==12.16.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-api-key==2.3.0
idna==3.4
isodate==0.6.1
kombu==5.2.4
msrest==0.6.21
oauthlib==3.2.2
prompt-toolkit==3.0.38
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
